requirements.txt
aiohttp
aiosignal
async-timeout
attrs
blivet==3.6.1
Bottleneck
brotlipy==0.7.0
cachetools
certifi==2022.12.7
cffi
charset-normalizer
click
colorama
cryptography
dataclasses
datasets
dialogflow==2.0.0
dill
filelock
Flask
frozenlist
fsspec
gmpy2
google-api-core==1.34.0
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-dialogflow==2.21.0
google-cloud-tasks==2.13.1
googleapis-common-protos
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.48.2
gunicorn==20.1.0
google-cloud-core
httplib2
huggingface-hub
idna
importlib-metadata
itsdangerous
Jinja2
joblib
MarkupSafe
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
multidict
multiprocess
networkx
numexpr
numpy
packaging
pandas==1.5.3
Pillow==9.4.0
proto-plus==1.22.2
protobuf==3.20.3
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser
pyOpenSSL
pyparsing
PySocks
python-dateutil
pytz
pyu2f
pyuwsgi==2.0.21
PyYAML
regex
requests
responses
rsa
sacremoses
six
sympy
tokenizers
torch==2.0.0
torchaudio==2.0.0
torchvision==0.15.0
tqdm
transformers
typing_extensions
uritemplate
urllib3
uWSGI
Werkzeug
xxhash
yarl
zipp
